Deduplicate checkbox type components code
